move convexhull_path() from make_ast_xy_list.py to cut_catalogs.py
